requirements.txt
# Core
numpy
opencv-python

# Window capture & control
pygetwindow
pyautogui
keyboard
mss
dxcam

# YOLO inference (OpenVINO)
openvino

# Gemma inference (OpenVINO)
optimum[openvino]
transformers
sentencepiece
protobuf

# Data collection
yt-dlp

# AI Video Labeling (Gemini)
google-genai
python-dotenv
